requirements.txt
Django==3.1.4
pip==20.3.1
django-filter==2.4.0
Pillow==8.1.0
